% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignKEGG.R
\name{assignKEGG}
\alias{assignKEGG}
\title{assignKEGG()}
\usage{
assignKEGG(input)
}
\arguments{
\item{input}{keggR BLAST table}
}
\value{
A keggR KO table
}
\description{
assignKEGG.
}
\examples{
assignKEGG(blast)
}
