% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignKEGG2.R
\name{assignKEGG2}
\alias{assignKEGG2}
\title{assignKEGG2()}
\usage{
assignKEGG2(input)
}
\arguments{
\item{input}{keggR HMM table}
}
\value{
A keggR KO table
}
\description{
assignKEGG2.
}
\examples{
assignKEGG2(blast)
}
