% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterKOtable.R
\name{filterKOtable}
\alias{filterKOtable}
\title{filterKOtable()}
\usage{
filterKOtable(input, seqs)
}
\arguments{
\item{input}{keggR KO table}
}
\value{
A filtered keggR KO table
}
\description{
filterKOtable.
}
\examples{
filterKOtable(blast)
}
