% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatKEGG.R
\name{formatKEGG}
\alias{formatKEGG}
\title{formatKEGG()}
\usage{
formatKEGG(path)
}
\arguments{
\item{path}{Path to the root of the KEGG directory}
}
\value{

}
\description{
Re-formats the KEGG auxiliary files for use with KEGG-tools.
}
\examples{
formatKEGG("/home/ipessi/KEGG")
}
