% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKOtable.R
\name{getKOtable}
\alias{getKOtable}
\title{getKOtable()}
\usage{
getKOtable(input)
}
\arguments{
\item{input}{keggR KO table}
}
\value{
A tibble
}
\description{
Extract KO table from a keggR KO table object.
}
\examples{
getKOtable(KOtable)
}
