% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{getSummary()}
\usage{
getSummary(input, mode = "pathways", level = "level1")
}
\arguments{
\item{input}{keggR summary table}
}
\value{
A tibble
}
\description{
getSummary.
}
\examples{
getSummary(summary)
}
