% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOtable2ANVIO.R
\name{KOtable2ANVIO}
\alias{KOtable2ANVIO}
\title{KOtable2ANVIO()}
\usage{
KOtable2ANVIO(input)
}
\arguments{
\item{input}{keggR KO table}
}
\value{
A tibble
}
\description{
KOtable2ANVIO
}
\examples{
KOtable2ANVIO(blast)
}
