% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadKEGG.R
\name{loadKEGG}
\alias{loadKEGG}
\title{loadKEGG()}
\usage{
loadKEGG(path)
}
\arguments{
\item{path}{Folder containing the KEGG auxiliary files}
}
\value{

}
\description{
Load the KEGG auxiliary files.
}
\examples{
loadKEGG("/home/igorspp/KEGG")
}
