% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSummaries.R
\name{mergeSummaries}
\alias{mergeSummaries}
\title{mergeSummaries()}
\usage{
mergeSummaries(input)
}
\arguments{
\item{input}{List of keggR summary tables}
}
\value{
A keggR summary table
}
\description{
mergeSummaries.
}
\examples{
mergeSummaries(list)
}
