% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBlast.R
\name{readBlast}
\alias{readBlast}
\title{readBlast()}
\usage{
readBlast(file)
}
\arguments{
\item{input}{BLAST file}
}
\value{
A keggR BLAST table
}
\description{
Read BLAST results in the tabular format (outfmt 6).
}
\examples{
readBlast("examples/input_data.txt")
}
