% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCustom.R
\name{readCustom}
\alias{readCustom}
\title{readCustom()}
\usage{
readCustom(file, colnames, evalue = 1)
}
\arguments{
\item{file}{Tab-delimited file}

\item{colnames}{Vector}

\item{evalue}{Double}
}
\value{
A keggR BLAST table
}
\description{
Read custom KEGG results.
}
\examples{
readCustom("examples/custom_input_data.txt")
}
