% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHMM.R
\name{readHMM}
\alias{readHMM}
\title{readHMM()}
\usage{
readHMM(file, e_value = FALSE)
}
\arguments{
\item{input}{HMM file}
}
\value{
A keggR HMM table
}
\description{
Read HMM results.
}
\examples{
readHMM("/examples/input_data.txt")
}
