% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMinpath.R
\name{runMinpath}
\alias{runMinpath}
\title{runMinpath()}
\usage{
runMinpath(input, path, debug = FALSE)
}
\arguments{
\item{input}{keggR KO table}
}
\value{
A keggR KO table
}
\description{
runMinpath.
}
\examples{
runMinpath(KOtable, "/home/igorspp/MinPath")
}
