% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseKEGG.R
\name{summariseKEGG}
\alias{summariseKEGG}
\title{summariseKEGG()}
\usage{
summariseKEGG(input)
}
\arguments{
\item{input}{keggR KO table}
}
\value{
A keggR summary table
}
\description{
summariseKEGG.
}
\examples{
summariseKEGG(KOtable)
}
