% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlastTable.R
\name{getBlastTable}
\alias{getBlastTable}
\title{getBlastTable()}
\usage{
getBlastTable(object)
}
\arguments{
\item{input}{keggR BLAST table}
}
\value{
A tibble
}
\description{
getBlastTable.
}
\examples{
getBlastTable(blast)
}
