% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBlast.R
\name{readBlast}
\alias{readBlast}
\title{readBlast()}
\usage{
readBlast(file)
}
\arguments{
\item{input}{BLAST file}
}
\value{
A keggR blast table
}
\description{
Read BLAST results.
}
\examples{
readBlast("/examples/input_data.txt")
}
